+++ Sources/ABAlbum/Views/AlbumGridCellView.swift
@@ -44,7 +44,7 @@ struct AlbumGridCellView: View {
         }
         .task {
             guard thumbnailImage == nil else { return }
-            async let stream = AlbumService.shared.asyncImage(from: asset, size: thumbnailSize, requestOptions: requestOptions)
+            async let stream = AlbumService.asyncImage(from: asset, size: thumbnailSize, requestOptions: requestOptions)
             do {
                 for try await image in await stream {
                     thumbnailImage = image